ebay-arbitrage-tool/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
playwright==1.49.0
beautifulsoup4==4.12.3
Pillow==11.0.0
anthropic==0.40.0
pydantic==2.10.3
lxml==5.3.0
websockets==13.1
aiofiles==24.1.0
